Clean up (pylint), additional static
Changed all tabs to spaces (becasue pylint)
Added a new static entry for the ansible control server (it would have been in the ec2 list, but this makes it easier to find).
Changed the "display_host" function to not echo the newline, so it only uses one newline when displaying the host entry.